Update patch set 4
Patch Set 4:

(3 comments)

I created a follow up due to my responsibility merging this. See Ic9fe89c7e3f984d3390260805e2480a0f8ffd0c9

> The patch still adds pretty complex code (almost a little scary, as parts still don't have obvious documentation) for very little advantage.

Personally I do not share this opinion when it comes to the complexity of the code. But that might be subjective. :-)

> Was the design team involved? Was this A/B tested? (At least on a small scale within the team?)

The change was reviewed with the PM and her decision was not involving the design team. Tests where done by comparing before and after results on several examples by PM and me.

Patch-set: 4