Update patch set 4
Patch Set 4:

Oh dear. I just fixed that bad documentation page. :-(

Passing nothing (technically undefined) as a parameter to mw.msg() is not the same as passing an empty string. But you are right, it's much better to follow the documentation. The best I found is in mw.jqueryMsg.HtmlEmitter.gender. It asks for the string "unknown".

As for the 50 users limit: I still vote for this hack, being aware of the consequences. As far as I can tell it's a rare edge-case to have more than 50 users in a page history, and the same time being able to have 50 of them on one screen (otherwise T197858 would have happened more often). And even if, what happens is that the 51st user (on the "less interesting" left side of the screen) will show a neutral label. I find this acceptable. It's certainly better than not having any genders. ;-)

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>