Update patch set 1
Patch Set 1:

Hey Peter and thanks for the patch. 

In general it looks sane, but still I would like to know if there's a Phabricator task or some more background on this? Is this related to a bug that turned up somewhere?

I would appreciate some more input on that. Thanks :-)

Patch-set: 1
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>