Update patch set 5
Patch Set 5:

(2 comments)

Ok I finally found the time to look at this ticket more in detail. In general it's good, but I guess two things could be improved.

Also when looking at the actual result of this, I had a quick talk with the PM and we decided on changing the numbers a bit. I will leave a comment on the ticket to be transparent there.

Patch-set: 5
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>