Update patch set 2
Patch Set 2: Code-Review-1

Nice, well done so far. But two things here that I see missing: 

When I get it right it says in the ticket, that the knobs should change to their brighter self when focused ( by tabbing ). You could probably expand the &:hover with an extra &:focus or so in the less file.

Also the diff view does not reload when moving is finished. Honestly I'm not entirely sure how this is handled best, but for now it would probably make sense to just reload the diff after each move, as soon as all other adjustments are made. - Probably somehow with refreshDiffView() - :-)

Patch-set: 2
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1