Update patch set 16
Patch Set 16:

(10 comments)

More comments! \o/ - So a lot of optimizations and some style things from my side.

We still have a general issue on how things work as soon as user is "marked". Removing and adding the class or getting rid of the mark again seems a bit tricky. 

I think it makes sense to have a class variable to save the "currently marked user". Most of the actions then could take this into account. e.g.:

- When I hover (mouseenter/leave) over a user's bubble and it is the currently marked user the hover has no effect.
- When I click on a user's bubble and it is the currently marked user all revisions will be unmarked and the currently marked user will be set to an empty sting.

This way you could get rid of the "toggle" action since we now only care for adds and removals and figure out whats need to be done by a combination of mouseevent and currently marked user.

I hope that's understandable :-D.

Patch-set: 16