Update patch set 3
Patch Set 3:

Okay! I am now applying the same style to the pointers as for when they are in hover state.  Also, added a blue outline to the knobs as some of the other elements have in `focus` state. Though, not sure, if that is the desired outcome. Also, now refreshing the diff view when the moving is finished.

Patch-set: 3
Reviewer: SrishtiSethi <5092@e9e9afe9-4712-486d-8885-f54b72dd1951>