Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

Cool, thanks for the improvements. I have some smaller comment in the code for optimizations and two further remarks:

The blue outlines are fine. - Ideally the bars around the focused pointer should be colored ( like when you hover a bar ) it might make the outline obsolete then. But this was not specified in the task I think and it could be done in a follow up later.

The other thing is, that the pointer loses focus when the diff is loaded. It would be better if this does not happen so that you can continue navigating. Probably you have to reset the focus via jquery and focus(); on the corresponding pointer.

Patch-set: 3
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1