Update patch set 5
Patch Set 5:

WoooW  \o/ nice !! Everything is good for me but
When focused on the yellow knobs, when they were adjacent to each other and press right arrow button again, I found the knob is not moving to the next as it's should not be, but the tooltip is showing the next revision info. I'm not sure whether it is intentional?
Same with the blue knob.

Patch-set: 5
Reviewer: Gopavasanth <5575@e9e9afe9-4712-486d-8885-f54b72dd1951>