Port expand Ruby selenium tests to NodeJS
These test do not add much value. Expanding the interface is an
essential step done in all other tests. Therefore the only test
ported is manually collapsing the interface.

Note that this is not completely cleaning up the step_definitions
from the Ruby tests. This will be done in a follow up after all
tests are ported.

Bug: T267198
Change-Id: I7beabc1a91f22dc8b7094318bf90684a3deb63f7