Port history Ruby selenium tests to NodeJS
The numbering in the content of the test revisions was slightly changed
to fit the data-pos attribute numbering in the revision tags.

Note that this is not completely cleaning up the step_definitions
from the Ruby tests. This will be done in a follow up after all
tests are ported.

Bug: T267201
Change-Id: Ifa25b881c6aa1e94a934532606c944a33f5648dd