selenium: Upgrade WebdriverIO to v7
Update npm packages: @wdio/* and webdriverio.

Also adding a fallback for the case, that the slider does
not open as expected.

It seems, that the update is causing some issues when
waiting for the module state. Waiting for the lazy module
in DiffPage.ready() should make sure, that the event to
trigger the opening of the slider is in place. This seems
to fail sometimes.

Bug: T274579
Change-Id: Icd597fc5aaa857879a8af310166867d433ba0678