Remove aria-hidden on RevisionSlider
I removed this aria-hidden, because the RevisionSlider as a whole
can be accessed with the focusable button that expands the UI.

Elements that are focusable should not be aria-hidden.
screenreaders in general will allow interaction with these or even
might show unpredictable behavior.

Screenreader and keyboard support for the feature is not good but
this aria-hidden is useless and wrong at the moment.

Bug: T339254
Change-Id: I8b8e3938cba701c821d4fd2ee3c2f90bf7610623