Restore highlightable-row CSS we accidentally removed
This was removed in I75ef7c3. Turns out it's needed for the
highlightable elements in the popups (e.g. when highlighting all
edits made by the same user). The text row in the popup gets a thin
gray border. This border needs some padding on the left.

The solution is to make the CSS selector more specific so it can
win over the problematic selector in Vector.

Bug: T341872
Change-Id: Ia029b92b5e049c60279b55177a62e03919dc55d8