Fix certain history events being triggered multiple times
See T349208 for an explanation. It looks like the SliderView.render
function was written with the assumption that it's only triggered
once on construction time. But since T139101 it's triggered again
for every window resize event. This adds the same event handlers
over and over again to existing elements that aren't affected by
the SliderView.render function.

This will probably become even worse with I49878fd (T336729).

Please test this carefully. I'm not 100% sure this is the best
possible fix.

Bug: T349208
Change-Id: Iba22924b660f2709c0680aa6fbeb0feba92cfa76