Revert "partly drop dependency from RevisionListView"
This is a manual revert of I06b8af8. Turns out the list of revisions
(and with that the maximum size calculated from the list of revisions)
changes during the lifetime of the view. It must be recalculated.
Maybe not as often. But this is an optimization for a later patch.

Change-Id: Id1dce5566a9041396bcf72df7e4936be0e53188a