Add fail-safe to SliderView.revisionsClickHandler
Turns out we can run into this failure situation even if everything
goes according to plan. It is possible to click just one pixel outside
of the valid range, e.g. on the very right corner right next to the
last bar. This is supposed to not do anything anyway, and correctly
doesn't do anything from the user's perspective. But it shows up as a
failure in the JavaScript console.

Bug: T352169
Change-Id: I12c9cce90970be36667ba1b721afd38a13a063c9