Merge a small piece of code duplication in SliderView
The method is only called from two places, and both do very similar
things before the call. This duplication can be merged.

One notable difference is that .scrollLeft() was possibly called
twice before. The first call was pointless anyway. The argument for
.scrollLeft() is absolute, not relative.

This code cleanup is motivated by T352169, but doesn't solve it.

Bug: T352169
Change-Id: I75e9ffc77ef6331f14e074921c78c28233e60840