Remove $dummy element after determineRtlScrollType call
This updates the code a little bit to be closer to the current
version of https://github.com/othree/jquery.rtl-scroll-type
without actually changing anything (for now).

I tried copy-pasting the current version to see if that helps with
T352169, but it doesn't make a difference.

The only actual change is that the dummy element is removed when
it's not needed any more.

Change-Id: Ibdd064e228fa9464a8652ce9a8a9ac388662f29d