Consolidate duplicate code updating slider line CSS
This code is updating the position and length of the yellow/blue
horizontal lines in the middle of the slider, left/right to the
yellow/blue circles.

1. setSliderLineCSS is called two times. A lot of the code before
is identical. Move all code that is identical into setSliderLineCSS.

2. The two single-use methods are very short now. Nothing but a
single function call. Merge the two calls into a single method.

Notice that `-0.5 * this.revisionWidth` is the same as
`-this.revisionWidth / 2`. I was able to make one disappear by
slowly transforming the code. It still does the same as before.

Change-Id: I9f9593bf4655fc8fa681fa8190db44dc8d6fc232