Pass through pointer events from pointer-containers
Hover and click events to highlight or change revisions need to
know the intended revision target. This did not work for the area
where the pointer containers overlap the underlying revisions.

For that reasons we implemented a calculation to get the revision
using the mouse positions. That implemetation seems to be faulty
at some points.

pointer-events: none allows us to pass though the mouse events so
that we're able to always rely on the revision containers as target.

On the pointers we still want to catch events to allow dragging.

Bug: T352169
Change-Id: Ie53a6ec3b7c458dc2f72e494829dfab80952b86f