Fix flipped left/right cursor keys in RTL mode
In RTL mode the slider is flipped and goes from right to left. The
revisions increase from right to left. The left cursor key needs to
be the one that increases the position.

Note there are three different "RTL scroll types" called "default",
"negative", and "reverse". This is only about CSS positions and
doesn't affect the cursor keys.

Change-Id: Ie747aa2572f2542b6c2c2176f817dd9a42b29f78