Remove confusing grab cursor when hovering ghosts
This line was added in 2017 via I9adbd47. I think this was a mistake.
What happens is that the grab cursor appears as part of the hover
effects (the yellow/blue "ghost" circles). But these can't be
dragged. Having the cursor there is confusing as it constantly
flickers between the click and drag cursors.

Am I missing something?

Change-Id: I611fee6f7199d68b61b103bcbb2011c88651883b